+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.4 - 4th August 2022
 - Correcting the name on the license
-- Updating the README setup instructions including the required Django settings 
+- Updating the README setup instructions including the required Django settings
+- Adding Exception handling into the Django Admin page so that the exceptions are caught and added to Django messages
 
 # 0.1.3 - 4th August 2022
 - License is now MIT